Migrate task events into taskevents.js
